controller/requirements.txt
# Google Cloud Platform
google-cloud-pubsub==2.33.0
google-cloud-storage==3.4.1

# HTTP client
requests==2.32.5

# Optional: Better logging and monitoring
python-json-logger==4.0.0
